Move `example_markers.js` to a better place
