+++ apps/web/web/static/js/leaflet/example_markers.js
